MarioMove/MarioSpecial: improve matches across both TUs
MarioMove:
- thinkParams 89.1% → 93.3% (nested ifs for short-circuit, remove skipCap)
- isSlipStart/canSlipJump/getSlideStopNormal/getSlideStopCatch ~96-97% (remove slipSet flag)
- gunExec 73.9% → 83.0% (PSVECScale/PSVECAdd instead of inline scale/add)
- checkGraffito 76.4% → 79.8% (getPollutionType return u32, isPolluted s32)
- setStatusToJumping 73.6% → 75.4% (type fixes, remove jmaSinShift locals)
- checkRideMovement 51.6% → 53.6% (eliminate existing local, swap if/else)
- checkWet: fix action bit mask 0x400 → 0x200

MarioSpecial:
- getCurrentPullParams 99.7% → 100% (reorder switch cases)
- moveRoof 49.7% → 80.6% (inline roofCommonEvents body)
- wireWait 60.3% → 76.0% (inline getOnWirePosAngle body)
- wireSWait 51.8% → 70.0% (inline getOnWirePosAngle body)
- doRoofMovingProcess 93.5% → 95.9% (expression reorder, swap nextPos.y/z)
- fencePunch 78.2% → 80.8% (add BOOL return type)
- wireMove 49.9% → 52.1% (TVec3 struct copy for dir)

+++ include/Map/PollutionManager.hpp
@@ -28,7 +28,7 @@ public:
 	void stamp(u16, f32 x, f32 y, f32 z, f32 range);
 	void clean(f32, f32, f32, f32);
 	void stampGround(u16, f32, f32, f32, f32);
-	u16 getPollutionType(f32, f32, f32) const;
+	u32 getPollutionType(f32, f32, f32) const;
 	u32 getPollutionDegree() const;
 	void isProhibit(f32, f32, f32) const;
 	bool isPolluted(f32, f32, f32) const;
